+++ packages/kofono-solid-form/CHANGELOG.md
@@ -1,5 +1,6 @@
 0.2.0
 - replace Kobalt by DaisyUI
+- moving to tsdown
 
 0.1.0 
 - proof of concept